requirements.txt
asgiref
chardet
crispy-tailwind
Django
redis
djangorestframework
django-crispy-forms
django-debug-toolbar
django-filter
django-widget-tweaks
django-templated-mail
django-redis
djoser
djangorestframework_simplejwt
Pillow
PyMySQL
python-dotenv
python-unsplash
requests
sqlparse
stripe
celery
flower
drf-nested-routers
whitenoise
gunicorn


